requirements.txt
Flask>=2.0
qrcode[pil]>=7.0
Pillow>=9.0
python-dotenv>=1.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2